Refactored file retrieval to separate file
