+++ benchmarking-results/txt-accuracy/nonorm_2025-06-17_140238.csv
@@ -1,0 +1,16 @@
+,gemini-2.5-flash,pytesseract
+kbaa-p100:dist_char,44,350
+kbaa-p100:doc_len,3854,3854
+kbaa-p100:cer_pct,1.1416709911779968,9.08147379346134
+kbaa-p100:wer_pct,7.177814029363784,57.09624796084829
+kbaa-p100:token_sort_ratio,99.05501618122977,96.26749611197512
+kbaa-p101:dist_char,60,347
+kbaa-p101:doc_len,3787,3787
+kbaa-p101:cer_pct,1.5843675732770002,9.162925798785318
+kbaa-p101:wer_pct,9.983361064891847,57.73710482529118
+kbaa-p101:token_sort_ratio,98.81141045958796,95.82180044813497
+__ALL__:dist_char,100,687
+__ALL__:doc_len,7642,7642
+__ALL__:cer_pct,1.308557969118032,8.989793247840879
+__ALL__:wer_pct,8.237232289950576,56.58978583196046
+__ALL__:token_sort_ratio,98.90842538728022,95.94250245017967
