+++ benchmarking-results/txt-accuracy/normalized_2025-06-17_140238.csv
@@ -1,0 +1,16 @@
+,gemini-2.5-flash,pytesseract
+kbaa-p100:dist_char,35,245
+kbaa-p100:doc_len,3491,3491
+kbaa-p100:cer_pct,1.0025780578630765,7.018046405041535
+kbaa-p100:wer_pct,5.747126436781609,40.229885057471265
+kbaa-p100:token_sort_ratio,99.09935668334523,97.63779527559055
+kbaa-p101:dist_char,51,251
+kbaa-p101:doc_len,3396,3396
+kbaa-p101:cer_pct,1.5017667844522968,7.391048292108363
+kbaa-p101:wer_pct,8.585858585858585,42.255892255892256
+kbaa-p101:token_sort_ratio,99.17574330291433,96.66127371672304
+__ALL__:dist_char,84,487
+__ALL__:doc_len,6888,6888
+__ALL__:cer_pct,1.2195121951219512,7.070267131242741
+__ALL__:wer_pct,6.982543640897755,40.482128013300084
+__ALL__:token_sort_ratio,99.06460735262127,97.15653561584216
