--- src/tools/file-retrieval.py
@@ -1,4 +1,0 @@
-"""
-Functions used for file retrieval.
-Script to get files for running accuracy tests.
-"""
